API
Rust optimization sweep: sync-only kernelctl, purpose-built Oil APK manager, release profile

- kernelctl: remove tokio, pure sync, 17-crate dep tree
- netd: u128→u64 unix ms timestamp
- glowfsctl: buffer write_zeroes via 4KB loop, write! hex_digest
- Oil: 28K→2.3K LOC, APK-only, ureq sync, no tracing/indicatif/rayon
- Cargo.toml: release profile (opt-level=z, lto=fat, strip, codegen-units=1)
- TODO.md: done/next-steps tracker

+++ system/netd/src/lib.rs
@@ -7,7 +7,7 @@ use serde::Serialize;
 
 #[derive(Clone, Debug, PartialEq, Serialize)]
 pub struct NetworkSnapshot {
-    pub generated_unix_ms: u128,
+    pub generated_unix_ms: u64,
     pub interfaces: Vec<NetworkInterface>,
 }
 
@@ -180,10 +180,10 @@ fn write_file(path: &Path, contents: &[u8]) -> io::Result<()> {
     fs::write(path, contents)
 }
 
-fn now_unix_ms() -> u128 {
+fn now_unix_ms() -> u64 {
     SystemTime::now()
         .duration_since(UNIX_EPOCH)
-        .map(|duration| duration.as_millis())
+        .map(|duration| duration.as_millis() as u64)
         .unwrap_or_default()
 }
